+++ src/CShuffle.h
@@ -1,0 +1,4 @@
+#include "SEutils.h"
+
+/*** R .Call Functions ***/
+void shuffleRInt(int *v, int *l){ shuffle(int, v, *l); }