+++ src/R_init_synextend.c
@@ -24,6 +24,9 @@
 // SynExtend header file
 #include "SynExtend.h"
 
+// Other header files for .C Routines
+#include "CShuffle.h"
+
 /*
  * -- REGISTRATION OF THE .Call ENTRY POINTS ---
  */
@@ -43,6 +46,9 @@ static const R_CallMethodDef callMethods[] = { // method call, pointer, num args
   {"calcDValue", (DL_FUNC) &calcDValue, 2},
   {"calcDRandValue", (DL_FUNC) &calcDRandValue, 4},
   {"calcDBrownValue", (DL_FUNC) &calcDBrownValue, 6},
+  {"pseudoRandomSample", (DL_FUNC) &pseudoRandomSample, 1},
+  {"randomProjection", (DL_FUNC) &randomProjection, 4},
+  {"seededPseudoRandomSample", (DL_FUNC) &seededPseudoRandomSample, 2},
   {NULL, NULL, 0}
 };
 
@@ -51,6 +57,8 @@ static const R_CallMethodDef callMethods[] = { // method call, pointer, num args
  */
 static const R_CMethodDef cMethods[] = {
   {"cleanupFxn", (DL_FUNC) &cleanupFxn, 0},
+  {"shuffleRInt", (DL_FUNC) &shuffleRInt, 2},
+  {"shuffleRDouble", (DL_FUNC) &shuffleRInt, 2},
   {NULL, NULL, 0}
 };