+++ NEWS.md
@@ -1,3 +1,13 @@
+# SynExtend 1.11.3
+* Internal bugfixes for JRF Distance--previous commit was incorrectly calculating values
+* Adds new `TreeDistance` predictor for `ProtWeaver`, incorporating all tree distance metrics; these metrics are bundled due to some backend optimizations that improve performance
+* Bugfixes for `PhyloDistance`
+* Adds Random Projection for `MirrorTree` predictor to solve memory problems and increase accuracy
+* New internal random number generator using xorshift, significantly faster than `sample()`
+* `HammingGL` changed to `CorrGL`, now uses Pearson's R weighted by p-value
+* Refactors internal predictors to reduce size of codebase and remove redundancies
+* Updates to documentation files
+
 # SynExtend 1.11.2
 * Adds KF Distance for trees
 * Adds Jaccard Robinson Foulds Distance for trees